done: minor refactoring - "fn date_time_from_timestamp_sec()" - made timestamp param u64

+++ src/worker/helper_functions.rs
@@ -17,12 +17,12 @@ pub fn strip_usd(pair: &(String, String)) -> Option<String> {
     }
 }
 
-pub fn date_time_from_timestamp_sec(timestamp_sec: i64) -> DateTime<Utc> {
-    let naive = NaiveDateTime::from_timestamp(timestamp_sec, 0);
+pub fn date_time_from_timestamp_sec(timestamp_sec: u64) -> DateTime<Utc> {
+    let naive = NaiveDateTime::from_timestamp(timestamp_sec as i64, 0);
 
     DateTime::from_utc(naive, Utc)
 }
 
-pub fn date_time_from_timestamp_millis(timestamp_millis: i64) -> DateTime<Utc> {
+pub fn date_time_from_timestamp_millis(timestamp_millis: u64) -> DateTime<Utc> {
     date_time_from_timestamp_sec(timestamp_millis / 1000)
 }
